refactor: change travelModal folder to rideModal
