+++ app/(tabs)/addRide.tsx
@@ -7,10 +7,10 @@ import Input from '@/components/input/Input'
 import { TextInputBlock } from '@/components/input/TextInput'
 import LoadingScreen from '@/components/LoadingScreen'
 import CustomDateTimePicker from '@/components/modal/CustomDatetimePicker'
-import AddRideLapsModal from '@/components/modal/travelModal/AddRideLapsModal'
-import EditRideDirectionModal from '@/components/modal/travelModal/EditRideDirectionModal'
-import EditRideRouteModal from '@/components/modal/travelModal/EditRideRouteModal'
-import EditRideStopModal from '@/components/modal/travelModal/EditRideStopModal'
+import AddRideLapsModal from '@/components/modal/rideModal/AddRideLapsModal'
+import EditRideDirectionModal from '@/components/modal/rideModal/EditRideDirectionModal'
+import EditRideRouteModal from '@/components/modal/rideModal/EditRideRouteModal'
+import EditRideStopModal from '@/components/modal/rideModal/EditRideStopModal'
 import { useDialog } from '@/context/DialogContext'
 import { useTheme } from '@/context/ThemeContext'
 import useDataOperations from '@/hooks/data/useDataOperations'
