+++ app/(tabs)/main/editRide.tsx
@@ -7,10 +7,10 @@ import Input from '@/components/input/Input'
 import { TextInputBlock } from '@/components/input/TextInput'
 import CustomDateTimePicker from '@/components/modal/CustomDatetimePicker'
 import { ManageableLap } from '@/components/modal/FlatlistPicker'
-import EditRideDirectionModal from '@/components/modal/travelModal/EditRideDirectionModal'
-import EditRideLapsModal from '@/components/modal/travelModal/EditRideLapsModal'
-import EditRideRouteModal from '@/components/modal/travelModal/EditRideRouteModal'
-import EditRideStopModal from '@/components/modal/travelModal/EditRideStopModal'
+import EditRideDirectionModal from '@/components/modal/rideModal/EditRideDirectionModal'
+import EditRideLapsModal from '@/components/modal/rideModal/EditRideLapsModal'
+import EditRideRouteModal from '@/components/modal/rideModal/EditRideRouteModal'
+import EditRideStopModal from '@/components/modal/rideModal/EditRideStopModal'
 import { useDialog } from '@/context/DialogContext'
 import { useModalContext } from '@/context/ModalContext'
 import { useTheme } from '@/context/ThemeContext'
