+++ app/(tabs)/main/estimate.tsx
@@ -4,9 +4,9 @@ import TypeButton from "@/components/button/TypeButton"
 import Container from "@/components/Container"
 import Divider from "@/components/Divider"
 import Input from "@/components/input/Input"
-import EditRideDirectionModal from "@/components/modal/travelModal/EditRideDirectionModal"
-import EditRideRouteModal from "@/components/modal/travelModal/EditRideRouteModal"
-import EditRideStopModal from "@/components/modal/travelModal/EditRideStopModal"
+import EditRideDirectionModal from "@/components/modal/rideModal/EditRideDirectionModal"
+import EditRideRouteModal from "@/components/modal/rideModal/EditRideRouteModal"
+import EditRideStopModal from "@/components/modal/rideModal/EditRideStopModal"
 import { JustifiedLabelValue } from "@/components/travel/IndividualTravelDetailCard"
 import { useTheme } from "@/context/ThemeContext"
 import useDirections from "@/hooks/data/useDirections"
